+++ tree-sitter-beans/queries/highlights.scm
@@ -43,6 +43,7 @@
     "Slice"
     "Atomic"
     "StoredCallback"
+    "CFunctionPtr"
     "Bytes"
     "File"
     "Dir"